chore(app/v2): organize auth function into dedicated files
